feat: add a class to update frames in a background thread
The `RTSPOpenCVStreamLoader` is a class that opens a new thread and continuously tries to fetch the new frame. It has a method called `load_frame()` which can be called from the main thread to fetch the most recent frame. It raises a custom exception `VideoCapError` if there was a problem with the stream. It's the responsibility of the caller to handle such an exception